Block face clipping polys almost done - need to do some refactoring first.

+++ src/main/java/grondag/adversity/niceblock/HeightController.java
@@ -65,7 +65,7 @@ public class HeightController extends ModelController implements ICollisionHandl
     @Override
     public long getCollisionKey(World worldIn, BlockPos pos, IBlockState state)
     {
-        return state.getValue(NiceBlock.META);
+        return (long) state.getValue(NiceBlock.META);
     }
 
     @Override
